move "create" view actions/constants to viewMap
