+++ createCard/actions/index.js
@@ -1,3 +1,2 @@
 export createCard from './createCard';
-export switchToCreate from './switchToCreate';
 export loadCards from './loadCards';
