+++ createCard/constants.js
@@ -1,4 +1,2 @@
 export const CREATE_NEW_CARD = 'CREATE_CARD_CREATE_NEW_CARD';
-export const SWITCH_TO_CREATE = 'SWITCH_TO_CREATE';
-export const CREATE_VIEW = 'createCards';
 export const LOAD_CARDS = 'CREATE_CARD_LOAD_CARDS';
