+++ viewDeck/index.js
@@ -2,8 +2,7 @@ import React from 'react';
 import {View, ScrollView, Text, StyleSheet} from 'react-native';
 import {Card, Button} from 'react-native-material-ui';
 import {connect} from 'react-redux';
-import viewHandler from '../viewHandler';
-import {Actions as CreateCardActions} from '../createCard';
+import {default as viewHandler,ViewActions as ViewActions} from '../viewMap';
 import {DECK_VIEW} from './constants';
 import {Actions as StudySessionActions} from '../studySession';
 import * as Actions from './actions';
@@ -58,7 +57,7 @@ const styles = StyleSheet.create({
 viewHandler.register(DECK_VIEW, connect(
     state => ({cards: state.cards || []}), 
     {
-        switchToCreate: CreateCardActions.switchToCreate, 
+        switchToCreate: ViewActions.switchToCreate, 
         switchToStudySession: StudySessionActions.switchToStudySession,
         ...Actions
     }
