+++ viewMap/actions/index.js
@@ -1,0 +1,1 @@
+export switchToCreate from './switchToCreate';
