+++ viewMap/actions/switchToCreate.js
