+++ viewMap/constants.js
@@ -1,0 +1,2 @@
+export const CREATE_VIEW = 'createCards';
+export const SWITCH_TO_CREATE = 'SWITCH_TO_CREATE';
